avance

+++ funciones/CAPI_r.py
@@ -5,7 +5,7 @@ def capi_r():
     datos_procesados = {}
     for indice, fila in df.iterrows():
         clave = fila[0]
-        valor = int(fila[1])
+        valor = float(fila[1])
         datos_procesados[clave] = valor
     return datos_procesados
