+++ funciones/L_i.py
@@ -16,6 +16,6 @@ def l_i():
     datos_procesados = {}
     for indice, fila in df.iterrows():
         clave = fila[0]
-        valor = int(fila[1])
+        valor = float(fila[1])
         datos_procesados[clave] = valor
     return datos_procesados
