+++ funciones/a_r.py
@@ -27,7 +27,7 @@ def ar():
     datos_procesados = {}
     for indice, fila in df.iterrows():
         clave = fila[0]
-        valor = int(fila[1])
+        valor = float(fila[1])
         datos_procesados[clave] = valor
     return datos_procesados
